+++ park_reviewer/src/App.css
@@ -1,9 +1,10 @@
 body {
   text-align: center;
-  background-color: #fefee3;
+  /* background-color: #fefee3; */
+  color: #34704e;
 }
 
-.test{
+.listWrapper{
   display: flex;
   flex-direction: row;
 }
