+++ park_reviewer/src/activityList.jsx
@@ -1,0 +1,34 @@
+import "./activityList.css"
+
+export default function ActivityList({location}){
+    return(
+        <div className="container">
+        <div className="activityList">
+            <h1>Top things to do in {location}</h1>
+            <div className="listData">
+            1. Example data
+            <br></br>
+            2. Example data
+            <br></br>
+            3. Example data
+            <br></br>
+            4. Example data
+            <br></br>
+            5. Example data
+            <br></br>
+            6. Example data
+            <br></br>
+            7. Example data
+            <br></br>
+            8. Example data
+            <br></br>
+            9. Example data
+            <br></br>
+            10. Example data
+            <br></br>
+        </div>
+        </div>  
+       
+        </div>
+    )
+}
