--- park_reviewer/src/footer.js
@@ -1,7 +1,0 @@
-import './headerFooter.css'
-
-export default function Footer(){
-    return(
-        <footer className="footer">Park Reviewer 2024</footer>
-    )
-}
