+++ park_reviewer/src/footer.jsx
@@ -1,0 +1,7 @@
+import './headerFooter.css'
+
+export default function Footer(){
+    return(
+        <footer className="footer"><p className='footerText'>Park Reviewer 2024</p></footer>
+    )
+}
