+++ park_reviewer/src/headerFooter.css
@@ -1,25 +1,41 @@
 .nav{
     width: 100%;
     height: 40px;
-    background-color: #4C956C;
+    background-color: #34704e;
     display: flex;
-    align-items: center;
+    flex-direction: row;
+    justify-content: space-between;
 }
    
 .navLink{
     display: inline-block;
     margin-right: 5px;
     text-decoration: none;
-    color: black;
+    color: white;
+    
+}
+
+.navDiv{
+    height: 80%;
+    width: auto;
+    margin: 5px;
+    align-content: center;
+    font-weight: 600;
 }
 
 .footer{
     width: 100%;
     height: 40px;
-    background-color: #4C956C;
+    background-color: #34704e;
     display:flex;
     align-items:center;
     position: fixed;
     left: 0;
     bottom: 0;
+    color: white;
+}
+
+.footerText{
+    margin-left: 5px;
+    font-weight: 600;
 }
