--- park_reviewer/src/navbar.js
@@ -1,10 +1,0 @@
-import './headerFooter.css'
-
-export default function NavBar(){
-    return(
-        <nav className="nav">
-                <a className="navLink" href="/" target="_blank">Home</a>
-                <a className="navLink" href="/" target="_blank">Login</a>
-        </nav>
-    )
-}
